requirements.txt
Flask==2.0.1
Pillow==8.2.0
tensorflow==2.6.0
numpy==1.19.5
opencv-python==4.5.3.20210927